Viva la Revolución! First Commit of Stand-alone Script
Starting from the previous project, where I was implementing the second-generation immigrants identification script together with household selection to retrieve debt and other relevant control variables, I have decided to create a stand-alone program. It is wiser to provide a base for the second-generation identification, such that everybody can use it. I left the track of all the previous commits, and I kept the branch where I firstly created the program that works on the mode of the nationality. I am planning to create a documentation soon.