Integration of Parents' Length of Stay and Age
I have integrated the calculation for the parents' length of stay in Germany using the spell dataset for the 2013-2015 Wave and the information about the parents' year of immigration. I use the biographical dataset and the linked parents to calculate their current age, considering the year of birth and death, and whether they are still alive or not.